moved handleAddProductsToCart to App Level
